+++ api/src/main/java/org/endeavourhealth/imapi/logic/service/FilerService.java
@@ -6,6 +6,7 @@ import org.endeavourhealth.imapi.filer.TTEntityFiler;
 import org.endeavourhealth.imapi.filer.TTFilerException;
 import org.endeavourhealth.imapi.filer.rdf4j.TTEntityFilerRdf4j;
 import org.endeavourhealth.imapi.filer.rdf4j.TTTransactionFiler;
+import org.endeavourhealth.imapi.logic.CachedObjectMapper;
 import org.endeavourhealth.imapi.logic.reasoner.SetExpander;
 import org.endeavourhealth.imapi.model.cdm.ProvActivity;
 import org.endeavourhealth.imapi.model.cdm.ProvAgent;
@@ -14,11 +15,14 @@ import org.endeavourhealth.imapi.model.search.EntityDocument;
 import org.endeavourhealth.imapi.model.tripletree.TTDocument;
 import org.endeavourhealth.imapi.model.tripletree.TTEntity;
 import org.endeavourhealth.imapi.model.tripletree.TTIriRef;
+import org.endeavourhealth.imapi.model.tripletree.TTValue;
 import org.endeavourhealth.imapi.vocabulary.GRAPH;
 import org.endeavourhealth.imapi.vocabulary.IM;
+import org.endeavourhealth.imapi.vocabulary.RDFS;
 import org.springframework.stereotype.Component;
 
-import static org.endeavourhealth.imapi.logic.service.EntityService.getBundle;
+import java.util.ArrayList;
+
 import static org.endeavourhealth.imapi.model.tripletree.TTIriRef.iri;
 
 @Component
@@ -72,7 +76,7 @@ public class FilerService {
     for (TTEntity entity : document.getEntities()) {
       TTEntity usedEntity = null;
       if (entityService.iriExists(entity.getIri())) {
-        usedEntity = getBundle(entity.getIri(), null).getEntity();
+        usedEntity = entityService.getBundle(entity.getIri(), null).getEntity();
       }
       ProvAgent agent = fileProvAgent(entity, agentName);
       TTEntity provUsedEntity = fileUsedEntity(usedEntity);
@@ -112,6 +116,79 @@ public class FilerService {
     } catch (Exception e) {
       throw new TTFilerException("Unable to file opensearch", e);
     }
+  }
 
+  public TTEntity createEntity(TTEntity entity, String agentName) throws TTFilerException, JsonProcessingException {
+    isValid(entity, "Create");
+    TTIriRef graph = iri(GRAPH.DISCOVERY);
+    entity.setCrud(iri(IM.ADD_QUADS)).setVersion(1);
+    fileEntity(entity, graph, agentName, null);
+    return entity;
+  }
+
+  public TTEntity updateEntity(TTEntity entity, String agentName) throws TTFilerException, JsonProcessingException {
+    isValid(entity, "Update");
+    TTIriRef graph = iri(GRAPH.DISCOVERY);
+    entity.setCrud(iri(IM.UPDATE_ALL));
+    TTEntity usedEntity = entityService.getBundle(entity.getIri(), null).getEntity();
+    entity.setVersion(usedEntity.getVersion() + 1);
+    fileEntity(entity, graph, agentName, usedEntity);
+    return entity;
+  }
+
+  public void isValid(TTEntity entity, String mode) throws TTFilerException, JsonProcessingException {
+    ArrayList<String> errorMessages = new ArrayList();
+    try (CachedObjectMapper om = new CachedObjectMapper()) {
+      if (Boolean.TRUE.equals(!isValidIri(entity))) errorMessages.add("Missing iri.");
+      if ("Create".equals(mode) && entityService.iriExists(entity.getIri())) errorMessages.add("Iri already exists.");
+      if ("Update".equals(mode) && !entityService.iriExists(entity.getIri())) errorMessages.add("Iri doesn't exists.");
+      if (Boolean.TRUE.equals(!isValidName(entity))) errorMessages.add("Name is invalid.");
+      if (Boolean.TRUE.equals(!isValidType(entity))) errorMessages.add("Types are invalid.");
+      if (Boolean.TRUE.equals(!isValidStatus(entity))) errorMessages.add("Status is invalid");
+      if (Boolean.TRUE.equals(!hasParents(entity))) errorMessages.add("Parents are invalid");
+      if (!errorMessages.isEmpty()) {
+        String errorsAsString = String.join(",", errorMessages);
+        throw new TTFilerException(mode + " entity errors: [" + errorsAsString + "] for entity " + om.writeValueAsString(entity));
+      }
+    }
+  }
+
+  private static Boolean isValidIri(TTEntity entity) {
+    if (null == entity.getIri()) return false;
+    return !"".equals(entity.getIri());
+  }
+
+  private static Boolean isValidName(TTEntity entity) {
+    if (null == entity.getName()) return false;
+    return !"".equals(entity.getName());
+  }
+
+  private static Boolean isValidType(TTEntity entity) {
+    if (null == entity.getType()) return false;
+    if (entity.getType().isEmpty()) return false;
+    return entity.getType().getElements().stream().allMatch(TTValue::isIriRef);
+  }
+
+  private static Boolean isValidStatus(TTEntity entity) {
+    if (null == entity.getStatus()) return false;
+    return entity.getStatus().isIriRef();
+  }
+
+  private static Boolean hasParents(TTEntity entity) {
+    if (null == entity.get(iri(IM.IS_A)) && null == entity.get(iri(IM.IS_CONTAINED_IN)) && null == entity.get(iri(RDFS.SUBCLASS_OF)) && null == entity.get(iri(IM.IS_SUBSET_OF)))
+      return false;
+    if (null != entity.get(iri(IM.IS_A)) && !entity.get(iri(IM.IS_A)).isEmpty()) {
+      if (!entity.get(iri(IM.IS_A)).getElements().stream().allMatch(TTValue::isIriRef)) return false;
+    }
+    if (null != entity.get(iri(IM.IS_CONTAINED_IN)) && !entity.get(iri(IM.IS_CONTAINED_IN)).isEmpty()) {
+      if (!entity.get(iri(IM.IS_CONTAINED_IN)).getElements().stream().allMatch(TTValue::isIriRef)) return false;
+    }
+    if (null != entity.get(iri(RDFS.SUBCLASS_OF)) && !entity.get(iri(RDFS.SUBCLASS_OF)).isEmpty()) {
+      if (!entity.get(iri(RDFS.SUBCLASS_OF)).getElements().stream().allMatch(TTValue::isIriRef)) return false;
+    }
+    if (null != entity.get(iri(IM.IS_SUBSET_OF)) && !entity.get(iri(IM.IS_SUBSET_OF)).isEmpty()) {
+      if (!entity.get(iri(IM.IS_SUBSET_OF)).getElements().stream().allMatch(TTValue::isIriRef)) return false;
+    }
+    return true;
   }
 }
